quick fix : Unnecessary boxing / unboxing

+++ src/main/java/tconstruct/modifiers/tools/ModButtertouch.java
@@ -66,7 +66,7 @@ public class ModButtertouch extends ModBoolean {
         boolean hasEnchant = false;
         while (iterator.hasNext()) {
             NBTTagCompound enchantTag = new NBTTagCompound();
-            index = ((Integer) iterator.next()).intValue();
+            index = (Integer) iterator.next();
             lvl = (Integer) enchantMap.get(index);
             if (index == enchant.effectId) {
                 hasEnchant = true;
